Mission accomplished.

+++ ProyectoInventarioOET/FormProductosLocales.aspx.cs
@@ -474,14 +474,14 @@ namespace ProyectoInventarioOET
                         datos[5] = "";
                         try
                         {
-                            if (Convert.ToInt32(datos[2].ToString()) <= Convert.ToInt32(datos[3].ToString()))
+                            if (Convert.ToDouble(datos[2].ToString()) <= Convert.ToDouble(datos[3].ToString()))
                             {
                                 datos[5] = "Debajo del mínimo";
 
                             }
                             else
                             {
-                                if (Convert.ToInt32(datos[2].ToString()) >= Convert.ToInt32(datos[4].ToString()))
+                                if (Convert.ToDouble(datos[2].ToString()) >= Convert.ToDouble(datos[4].ToString()))
                                 {
                                     datos[5] = "Encima del máximo";
                                 }
